Split building app views
