+++ buildings/views/units.py
@@ -2,135 +2,20 @@ from django.contrib import messages
 from django.db import transaction
 from django.shortcuts import get_object_or_404
 from django.shortcuts import redirect
-from django.urls import reverse
 from django.utils.translation import ugettext as _
-from django.views.generic import CreateView
 from django.views.generic import DetailView
 from django.views.generic import ListView
 from django.views.generic import TemplateView
-from django.views.generic import UpdateView
-from django.views.generic.edit import DeleteView
 
-from .forms import BuildingForm
-from .forms import LeaseholderFormSet
-from .forms import OwnerFormSet
-from .forms import UnitForm
-from .forms import ParkingLotForm
-from .models import Building
-from .models import BuildingMembership
-from .models import Leaseholder
-from .models import Owner
-from .models import ParkingLot
-from .models import Unit
-from .permissions import BuildingPermissions
 from app.mixins import CustomUserMixin
-
-
-class BuildingFormView(CustomUserMixin, CreateView):
-    """
-    Form view to create a new Building or condo.
-    A membership with administrator role is created
-    to the authenticated user.
-    """
-    model = Building
-    form_class = BuildingForm
-    template_name = 'buildings/building_form.html'
-
-    def test_func(self):
-        return BuildingPermissions.can_create_building(
-            user=self.request.user,
-        )
-
-    @transaction.atomic
-    def form_valid(self, form):
-        building = form.save(commit=False)
-        building.created_by = self.request.user
-        building.save()
-
-        BuildingMembership.objects.create(
-            user=self.request.user,
-            building=building,
-            is_administrator=True,
-            is_active=True,
-        )
-
-        messages.success(
-            self.request,
-            _('Copropiedad creada exitosamente.')
-        )
-
-        return redirect(building.get_absolute_url())
-
-
-class BuildingDetailView(CustomUserMixin, DetailView):
-    """
-    Detail view of a building or condo.
-    """
-    model = Building
-    template_name = 'buildings/administrative/building_detail.html'
-
-    def get_object(self, queryset=None):
-        return get_object_or_404(
-            Building,
-            pk=self.kwargs['pk'],
-        )
-
-    def test_func(self):
-        return BuildingPermissions.can_view_building_detail(
-            user=self.request.user,
-            building=self.get_object(),
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['can_edit_building'] = BuildingPermissions.can_edit_building(
-            user=self.request.user,
-            building=self.get_object(),
-        )
-
-        context['active_general'] = True
-
-        return context
-
-
-class BuildingUpdateView(CustomUserMixin, UpdateView):
-    """
-    Form view to update basic information about a building
-    or condo.
-    """
-    model = Building
-    form_class = BuildingForm
-    template_name = 'buildings/building_form.html'
-
-    def test_func(self):
-        return BuildingPermissions.can_edit_building(
-            user=self.request.user,
-            building=self.get_object(),
-        )
-
-    def get_object(self, queryset=None):
-        building = get_object_or_404(
-            Building,
-            pk=self.kwargs['pk'],
-        )
-
-        return building
-
-    def get_success_url(self):
-        return reverse(
-            'buildings:building_detail', args=[self.kwargs['pk']]
-        )
-
-    @transaction.atomic
-    def form_valid(self, form):
-        form.save()
-
-        messages.success(
-            self.request,
-            _('Copropiedad actualizada correctamente.'),
-        )
-
-        return super().form_valid(form)
+from buildings.forms import LeaseholderFormSet
+from buildings.forms import OwnerFormSet
+from buildings.forms import UnitForm
+from buildings.models import Building
+from buildings.models import Leaseholder
+from buildings.models import Owner
+from buildings.models import Unit
+from buildings.permissions import BuildingPermissions
 
 
 class UnitsListView(CustomUserMixin, ListView):
@@ -455,141 +340,5 @@ class UnitDetailView(CustomUserMixin, DetailView):
             user=self.request.user,
             building=self.get_object().building,
         )
-
-        return context
-
-
-class ParkingLotFormView(CustomUserMixin, CreateView):
-    """
-    Form view to create a new parking lot of a unit.
-    """
-    model = ParkingLot
-    form_class = ParkingLotForm
-    template_name = 'buildings/administrative/parkinglot_form.html'
-
-    def test_func(self):
-        return BuildingPermissions.can_edit_unit(
-            user=self.request.user,
-            building=self.get_object().building,
-        )
-
-    def get_object(self, queryset=None):
-        return get_object_or_404(
-            Unit,
-            building_id=self.kwargs['b_pk'],
-            pk=self.kwargs['u_pk'],
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['unit'] = self.get_object()
-        context['building'] = self.get_object().building
-        context['active_units'] = True
-
-        return context
-
-    @transaction.atomic
-    def form_valid(self, form):
-        # Get unit instance.
-        unit = self.get_object()
-        # Create parking lot object.
-        parking_lot = form.save(commit=False)
-        parking_lot.unit = self.get_object()
-        parking_lot.save()
-
-        messages.success(
-            self.request,
-            _('Parqueadero creado exitosamente.')
-        )
-
-        return redirect(unit.get_absolute_url())
-
-
-class ParkingLotUpdateView(CustomUserMixin, UpdateView):
-    """
-    Form view to update information about a parking lot.
-    """
-    model = ParkingLot
-    form_class = ParkingLotForm
-    template_name = 'buildings/administrative/parkinglot_form.html'
-
-    def test_func(self):
-        return BuildingPermissions.can_edit_unit(
-            user=self.request.user,
-            building=self.get_object().unit.building,
-        )
-
-    def get_object(self, queryset=None):
-        # Get parking lot object.
-        return get_object_or_404(
-            ParkingLot,
-            unit_id=self.kwargs['u_pk'],
-            pk=self.kwargs['p_pk'],
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['unit'] = self.get_object().unit
-        context['building'] = self.get_object().unit.building
-        context['active_units'] = True
-        context['parking_lot_update'] = True
-
-        return context
-
-    def get_success_url(self):
-        # Reverse to unit detail.
-        return reverse(
-            'buildings:unit_detail',
-            args=[self.kwargs['b_pk'], self.kwargs['u_pk']]
-        )
-
-    @transaction.atomic
-    def form_valid(self, form):
-        # Update parking lot object.
-        form.save()
-
-        messages.success(
-            self.request,
-            _('Parqueadero actualizado exitosamente.'),
-        )
-
-        return super().form_valid(form)
-
-
-class ParkingLotDeleteView(CustomUserMixin, DeleteView):
-    """
-    Parking lot delete view. Users is redirected to a view
-    in which they will be asked about confirmation for
-    delete a parking lot definitely.
-    """
-    model = ParkingLot
-    template_name = 'buildings/administrative/parking_lot_delete_confirm.html'
-
-    def test_func(self):
-        return BuildingPermissions.can_edit_unit(
-            user=self.request.user,
-            building=self.get_object().unit.building,
-        )
-
-    def get_object(self, queryset=None):
-        # Get parking lot object.
-        return get_object_or_404(
-            ParkingLot,
-            unit_id=self.kwargs['u_pk'],
-            pk=self.kwargs['p_pk'],
-        )
-
-    def get_success_url(self):
-        # Reverse to unit detail.
-        return reverse(
-            'buildings:unit_detail',
-            args=[self.kwargs['b_pk'], self.kwargs['u_pk']]
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['unit'] = self.get_object().unit
-        context['active_units'] = True
-        context['building'] = self.get_object().unit.building
 
         return context
